Minor fixes

+++ src/middlewared/middlewared/plugins/device_/device_info_freebsd.py
@@ -85,7 +85,7 @@ class DeviceService(Service, DeviceInfoBase):
         if disk['serial'] and disk['lunid']:
             disk['serial_lunid'] = f'{disk["serial"]}_{disk["lunid"]}'
         if disk['size'] and disk['sectorsize']:
-            disk['blocks'] = disk['size'] / disk['sectorsize']
+            disk['blocks'] = int(disk['size'] / disk['sectorsize'])
 
         return disk
